GUI update and code changes
Added some simple things to the UI as placeholders for now to get a better idea on what to add in the future. Refactored some of the code so that the UIHandler is in control of button click events, not the state machine. The state machine should be independent of anything having to do with the UI.